pyproject.toml
[project]
name = "structorex"
version = "0.2.0"
description = "Project structure generator with file content inspection"
authors = [{ name = "Maxim Shushanikov", email = "max.shushanikov@gmail.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
    "pathspec>=0.11.0",
    "PyYAML>=6.0",
    "tomli>=2.0.0; python_version < '3.11'",
    "tqdm>=4.65.0",
    "charset-normalizer>=3.1.0"
]
keywords = ["project-structure", "scaffolding", "inspection", "generator"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Environment :: Console",
  "Topic :: Software Development :: Code Generators"
]

[project.scripts]
structorex = "structorex.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pytest-cov>=4.0",
  "pytest-mock>=3.0",
  "flake8>=6.0",
  "build>=1.0.0",
  "twine>=4.0.0",
  "mkdocs>=1.5.0",
  "mkdocs-material>=9.0.0",
  "mkdocstrings[python]>=0.24.0",
  "mkdocs-static-i18n>=0.56"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true
license-files = []

[tool.setuptools.packages.find]
where = ["src"]
include = ["structorex*"]
exclude = ["site*", "tests*"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
pythonpath = ["."]
testpaths = ["tests"]